Revert "Verify user"
