+++ package.json
@@ -18,8 +18,6 @@
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
-    "react-hook-form": "^7.50.1",
-    "react-icons": "^5.0.1",
     "zustand": "^4.5.0"
   },
   "devDependencies": {
