+++ src/app/(main)/layout.tsx
@@ -6,7 +6,6 @@ import MainNavbar from "@/components/main/MainNavbar";
 import Footer from "@/components/main/Footer";
 import { MantineProvider } from "@mantine/core";
 import getUserSession from "@/libs/actions/getSession";
-import UserNavbar from "@/components/main/UserNavbar";
 export const lora = Lora({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
@@ -27,7 +26,7 @@ export default async function RootLayout({
       <html lang="en">
         <body className={lora.className}>
           <MantineProvider>
-            <UserNavbar />
+            <MainNavbar />
             {children}
           </MantineProvider>
         </body>
