--- src/app/(verify-user)/layout.tsx
@@ -1,28 +1,0 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "@/styles/globals.css";
-import "@mantine/core/styles.css";
-import VerifyNavbar from "@/components/main/VerifyNavbar";
-export const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: {
-    default: "TechVibe",
-    template: "%s - TechVibe",
-  },
-  description: "Come and read Our Blog",
-};
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <VerifyNavbar />
-        {children}
-      </body>
-    </html>
-  );
-}
