--- src/types/navbar/navbar-tpye.d.ts
@@ -1,6 +1,0 @@
-export type NoticardProps = {
-  type: string;
-  title: string;
-  content: string;
-  status: string;
-};
