+++ src/types/user/user.d.ts
@@ -1,19 +1,4 @@
 export type userProps = {
-  user_id?: string;
-  user_email?: string;
-  user_fullname?: string;
-  user_provider?: string;
-  created_at?: string;
-  updated_at?: string;
-  user_verify?: boolean;
-};
-
-export type userSocialProps = {
-  user_social: {
-    user_social_id: string;
-    user_social_facebook: string;
-    user_social_github: string;
-    user_social_twitter: string;
-    user_id: string;
-  };
+  firstName: string;
+  role: string;
 };
